+++ CHANGELOG.md
@@ -7,6 +7,7 @@
  + method `get_me` to class `Site`;
  + method `get_my_inbox` to class `Site`;
  + method `get_my_unread_inbox` to class `Site`;
+ + method `get_unanswered_questions_on_my_tags` to class `Site`;
  + method `post` to class `Site`
 
 ### Changed: